fix(ios): split ParentDashboardView to satisfy SwiftLint type_body_length
Extract dashboard sections into ParentDashboardSections.swift so the
main view struct stays under the 350-line SwiftLint limit that caused
CI — iOS to fail.
